Move UrlChecker to its own namespace.
